+++ hadoop-hdds/container-service/src/main/java/org/apache/hadoop/ozone/container/keyvalue/interfaces/BlockManager.java
@@ -90,7 +90,7 @@ public interface BlockManager {
   long getCommittedBlockLength(Container container, BlockID blockID)
       throws IOException;
 
-  long getDefaultReadBufferCapacity();
+  int getDefaultReadBufferCapacity();
 
   /**
    * Shutdown ContainerManager.
